# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TinyForgeAI Inference API",
    "version": "0.1.0",
    "description": "REST API for TinyForgeAI inference services. Train tiny language models from your data and deploy them as microservices.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "contact": {
      "name": "TinyForgeAI",
      "url": "https://github.com/anthropics/TinyForgeAI"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/predict": {
      "post": {
        "summary": "Perform inference",
        "description": "Send input text to the model and receive a prediction with confidence score.",
        "operationId": "predict",
        "tags": ["Inference"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              },
              "examples": {
                "simple": {
                  "summary": "Simple text input",
                  "value": {
                    "input": "hello world"
                  }
                },
                "question": {
                  "summary": "Question input",
                  "value": {
                    "input": "What is your refund policy?"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful prediction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                },
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "output": "dlrow olleh",
                      "confidence": 0.75
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PredictRequest": {
        "type": "object",
        "required": ["input"],
        "properties": {
          "input": {
            "type": "string",
            "description": "The input text for inference",
            "example": "hello world",
            "minLength": 1
          },
          "metadata": {
            "type": "object",
            "description": "Optional metadata to include with the request",
            "additionalProperties": true
          }
        }
      },
      "PredictResponse": {
        "type": "object",
        "properties": {
          "output": {
            "type": "string",
            "description": "The model's output prediction"
          },
          "confidence": {
            "type": "number",
            "format": "float",
            "description": "Confidence score between 0 and 1",
            "minimum": 0,
            "maximum": 1
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "loc": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "Location of the error"
                },
                "msg": {
                  "type": "string",
                  "description": "Error message"
                },
                "type": {
                  "type": "string",
                  "description": "Error type"
                }
              }
            }
          }
        }
      }
    }
  },
...
}